@@ -9,7 +9,7 @@
 
 import React, { Component } from 'react';
 import withStyles from 'isomorphic-style-loader/lib/withStyles';
-import LazyLoad from '@components/common/Lazyload';
+import LazyLoadBeta from '@components/common/LazyloadBeta';
 import Link from '../Link';
 // import { Link } from 'react-router-dom';
 
@@ -17,41 +17,41 @@ import styles from './Navigation.css';
 
 class Navigation extends Component {
   handleNavigation = e => {
-      e.preventDefault();
-      const { onClick } = this.props;
-      onClick(e.currentTarget.id);
+    e.preventDefault();
+    const { onClick } = this.props;
+    onClick(e.currentTarget.id);
   };
 
   render() {
-      const { isDark = 1, playlists = [] } = this.props,
-          color = isDark ? 'black' : 'white';
-      return (
-          <div className={styles.navigation__wrapper} style={{ color }}>
-              {playlists &&
+    const { isDark = 1, playlists = [] } = this.props,
+      color = isDark ? 'black' : 'white';
+    return (
+      <div className={styles.navigation__wrapper} style={{ color }}>
+        {playlists &&
           playlists.map(
-              ({ id, isActive, title }) => (
-                  <Link
-                      key={id}
-                      id={id}
-                      to="/"
-                      onClick={this.handleNavigation}
-                      className={[
-                          styles.navigation__links,
-                          isActive ? styles.isActive : '',
-                      ].join(' ')}
-                      style={{ color }}
-                  >
-                      {isActive && (
-                          <LazyLoad>
-                              <hr style={{ borderBottom: `1px solid ${color}` }} />
-                          </LazyLoad>
-                      )}
-                      {title.toUpperCase()}
-                  </Link>
-              ),
+            ({ id, isActive, title }) => (
+              <Link
+                key={id}
+                id={id}
+                to="/"
+                onClick={this.handleNavigation}
+                className={[
+                  styles.navigation__links,
+                  isActive ? styles.isActive : '',
+                ].join(' ')}
+                style={{ color }}
+              >
+                {isActive && (
+                  <LazyLoadBeta>
+                    <hr style={{ borderBottom: `1px solid ${color}` }} />
+                  </LazyLoadBeta>
+                )}
+                {title.toUpperCase()}
+              </Link>
+            ),
           )}
-          </div>
-      );
+      </div>
+    );
   }
 }
 
