@@ -1,3 +1,5 @@
-const GET_HISTORY = 'history/GET_HISTORY';
+const GET_HISTORY_LOADING = 'history/GET_HISTORY_LOADING';
+const GET_HISTORY_SUCCESS = 'history/GET_HISTORY_SUCCESS';
+const GET_HISTORY_ERROR = 'history/GET_HISTORY_ERROR';
 
-export { GET_HISTORY };
+export { GET_HISTORY_LOADING, GET_HISTORY_SUCCESS, GET_HISTORY_ERROR };
