@@ -277,9 +277,9 @@ class Home extends Component {
                                     >
                                       <LazyLoadBeta src={layer3}
                                         containerClassName={styles.home__parallax_layer_3_info}
-                                        style={{ color: isDark ? "black" : "white" }}
                                       >
-                                        <div className={styles.home__parallax_layer_3_detail}>
+                                        <div className={styles.home__parallax_layer_3_detail}
+                                          style={{ color: isDark ? "black" : "white" }}>
                                           <h4
                                             className={styles.home__parallax_layer_3_title}
                                           >
@@ -290,7 +290,6 @@ class Home extends Component {
                                             <Link to="/movie" className={styles.home__see_more}>➪see movie</Link>
                                           </p>
                                         </div>
-
                                       </LazyLoadBeta>
                                     </Parallax>
                                     <Parallax
